break up mega file chart_operations for ease of maintenance
